pyproject.toml
[project]
name = "photo-tools"
version = "0.1.0"
description = "Auto-tag photos using RAM++, CLIP landmarks, PaddleOCR, GPS geocoding, and EXIF metadata"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow-heif>=1.3.0",
    "pyyaml>=6.0",
    "requests>=2.33.1",
    "numpy>=2.0",
    "torch>=2.0",
    "torchvision>=0.15",
    "open-clip-torch>=2.24.0",
    "ram",
    "transformers<=4.55.4",
    "huggingface-hub>=0.20.0",
    "faiss-cpu>=1.7.0",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "scipy",
    "timm==0.4.12",
    "fairscale==0.4.4",
    "Pillow",
    "titlecase>=2.4",
    "rich>=15.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]

[build-system]
requires = ["uv_build>=0.7.2,<0.8"]
build-backend = "uv_build"

[project.scripts]
photo-tools = "photo_tools.cli:main"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = {index = "pytorch-cpu"}
torchvision = {index = "pytorch-cpu"}
ram = {git = "https://github.com/xinyu1205/recognize-anything.git"}
